movieDb/requirements.txt
django==2.2.2
djangorestframework
requests
